+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 1.0.0
+## 0.0.1
 
 - Initial version.
+- Build and sign ios application.
+- Disk infrastructure manager.